Initial patches/workarounds
Custom changes to script.pseudotv.live for problems I have encountered
over the past few releases.
--EPGWindow.py--
1. VideowWindow patch works, but throws some exception, causing it to
disappear after a second, upon entering the EPG.  This modification
keeps VideoWindow running, but does not address the cause of the
exception.
--Overlay.py
1. On my system, when changing a channel, sometimes the video will
start from the beginning even though it should be in the middle of a
show or movie.  This is most annoying when watching a movie, changing
the channel briefly, then changing back and the movie restarts.  I've
traced it back to self.Player.seektime() being called before Kodi is
playing the video it just changed to.  Here we call a loop to wait until
xmbc.Player().getTime() returns a valid number to show that the video
has startedplaying. Drawback: May cause a few ms delay when changing a
channel if you do not suffer from the problem I have been having.
2.  When pausing, PTVL shows the channel background. I prefer the
older behavior where it just paused the video and showed the still
frame.  This change just comments out the background screen display when
pausing.